+++ python/metatensor_torch/metatensor/torch/operations.py
@@ -2,8 +2,9 @@ import importlib
 import os
 import sys
 
-import metatensor.operations
 import torch
+
+import metatensor.operations
 from metatensor.torch import Labels, TensorBlock, TensorMap